refactor(api): déplacer le modèle serializeComment vers shantytownCommentModel refers to https://trello.com/c/evsOelCU/2717
- déplace serializeComment de models/shantytownModel/_common vers models/shantytownCommentModel
- exporte serializeComment depuis le barrel shantytownCommentModel
- supprime l'export depuis le barrel shantytownModel
- met à jour les imports dans findByShantytown, findOne, getHistory et serializeComment
